Create Tingkat and Kelas model.
Also create a pre save signal for the unique together of all attribute.
If you create a Kelas with the same attribute as an existing one, that signal
will raise a ValidationError. Also changing the pre save signal on KKM
to behave exactly like this one, rather than deleting all of the existing record.